chore(engine): standardize Ruff lint ruleset and clean up violations
Expand the engine Ruff config beyond E/F/I/UP/B/SIM to a broader
"clean code" set — C4, RET, PIE, PTH, RUF, TID, and Pylint (PL, which
includes PLC0415 "import should be at top-level"). Ruff reimplements
pylint's checks, so no separate pylint tool is needed.

Lint config & tooling:
- Align ruff versions: bump ruff-pre-commit v0.11.4 -> v0.15.9 to match
  the locally-resolved ruff. In 0.11.4 PLC0415 was still preview (not
  enforced), so the hook and local dev disagreed and stripped/required
  `# noqa: PLC0415` inconsistently. Now both treat it as stable.
- per-file-ignore the intentional lazy imports: telemetry/* and
  events/services.py (opentelemetry), cli/* (fast startup),
  server/app.py (router wiring), llm_providers/services.py (anthropic/
  openai/urllib per-provider SDKs).
- RUF012 ignored for starlette-admin ModelViews (framework API);
  PLC0415/RUF059 relaxed under tests/.

Code cleanups:
- Hoist genuinely-unnecessary in-function imports to module top
  (manager._auto_introspect, store.soft_delete, admin.views.mount_admin,
  routes test_connection/introspect_connection); drop one redundant
  re-import of GraphUnavailableError. Confirmed no circular imports.
- RUF006: add GraphConnectionManager._spawn() to retain strong refs to
  fire-and-forget connect tasks so the loop can't GC them mid-flight.
- Mechanical fixes: RET504, C4xx, PLR1714, RUF046, PLW0108, PIE790,
  .items() rewrites in modeller/versioner.py, sorted __all__/__slots__,
  and dropped now-unused noqa comments (BLE001/S105/S310/N807/F401).

Also includes a pre-existing working-tree edit to _auto_introspect:
schema name now derives from the parent Graph (parent.name) with a
fallback to the connection URI, instead of graph.name.

+++ engine/src/invana/server/routes/query.py
@@ -170,7 +170,7 @@ def _build_query_response(graph_response: GraphResponse, query_language: QueryLa
             query_language=query_language.value,
             data=graph_response,
             rows=None,
-            execution_time_ms=int(round(graph_response.metadata.duration_ms)),
+            execution_time_ms=round(graph_response.metadata.duration_ms),
             row_count=len(graph_response.nodes) + len(graph_response.edges),
         )
     return QueryResponse(
@@ -178,6 +178,6 @@ def _build_query_response(graph_response: GraphResponse, query_language: QueryLa
         query_language=query_language.value,
         data=None,
         rows=graph_response.records,
-        execution_time_ms=int(round(graph_response.metadata.duration_ms)),
+        execution_time_ms=round(graph_response.metadata.duration_ms),
         row_count=len(graph_response.records),
     )
